Add ReaderFeaturedImage block
