+++ assets/stylesheets/_components.scss
@@ -35,6 +35,7 @@
 @import 'blocks/post-status/style';
 @import 'blocks/reader-avatar/style';
 @import 'blocks/reader-feed-header/style';
+@import 'blocks/reader-featured-image/style';
 @import 'blocks/reader-full-post/style';
 @import 'blocks/reader-post-actions/style';
 @import 'blocks/reader-post-card/style';
