+++ assets/stylesheets/shared/functions/_z-index.scss
@@ -44,7 +44,7 @@ $z-layers: (
 		'.reader__featured-post-title': 1,
 		'.search-stream__input-card': 0,
 		'.reader-search-card': 0,
-		'.reader-post-card__featured-image': 0,
+		'.reader-featured-image': 0,
 		'.reader-post-card__title': 0,
 		'.reader-list-gap__button': 1,
 		'.reader-following-search': 1,
