+++ client/blocks/reader-featured-image/README.md
@@ -1,0 +1,17 @@
+# Reader Featured Image
+
+Displays a featured image for a Reader post card.
+
+## Example
+
+The `ReaderFeaturedImage` component can be used in much the same way that you would use an `<ExternalLink>` component. The link wraps whatever is placed between the ReaderFeaturedImage elements.
+
+```html
+<ReaderFeaturedImage imageUrl={ imageUrl } href={ href }>[button]</ReaderFeaturedImage>
+```
+
+## Props
+
+- `imageUrl`: Image URL
+- `href`: The URL to link the featured image to
+- `onClick`: Function to execute when image is clicked
