+++ client/blocks/reader-featured-image/style.scss
@@ -1,0 +1,3 @@
+.reader-featured-image {
+	border: 1px solid lighten( $gray, 20% );
+}
